requirements.txt
paho-mqtt==1.4.0
pika==1.1.0